refactor: rename AnkiServiceRefactored to AnkiService - Renamed AnkiServiceRefactored.ts to AnkiService.ts - Renamed original AnkiService.ts to AnkiServiceLegacy.ts for preservation - Updated all imports and class references in routes and tests - Updated test file name and descriptions to match new service name - All linting and compilation checks pass
